materialize: move to api/ to be closer to its user
